refactor: combine form helpers in a single file
